requirement.txt
asgiref==3.3.1
autopep8==1.5.4
cffi==1.14.4
cryptography==3.2.1
Django==3.1.4
django-debug-toolbar==3.2
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
